KPMP-3883: Added new code to get the explorer dataset information

+++ src/main/java/org/kpmp/dataSummary/DataSummaryRepository.java
@@ -29,8 +29,8 @@ public interface DataSummaryRepository extends CrudRepository<DataSummaryValue, 
 	Long getParticipantSummaryLinkCount(@Param("data_type") String data_type);
 
 	@Query(value = "SELECT count(*) from sv_file_v WHERE data_type= :data_type AND redcap_id= :redcap_id", nativeQuery = true)
-	Long getParticipantSvFileDataTypeCount(@Param("redcap_id") String redcapId, @Param("data_type") String dataType);
+	Integer getParticipantSvFileDataTypeCount(@Param("redcap_id") String redcapId, @Param("data_type") String dataType);
 
 	@Query(value = "SELECT count(*) from sv_link_v WHERE data_type= :data_type AND redcap_id= :redcap_id", nativeQuery = true)
-	Long getParticipantSvLinkDataTypeCount(@Param("redcap_id") String redcapId, @Param("data_type") String dataType);
+	Integer getParticipantSvLinkDataTypeCount(@Param("redcap_id") String redcapId, @Param("data_type") String dataType);
 }
